Revert "Implement UnifiedCallbackCoordinator"
