+++ core/app_factory.py
@@ -8,7 +8,7 @@ import os
 from typing import Optional, Any
 import dash_bootstrap_components as dbc
 from dash import html, dcc, Input, Output, callback
-from core.unified_callback_coordinator import UnifiedCallbackCoordinator
+from core.callback_manager import CallbackManager
 import pandas as pd
 
 # ✅ FIXED IMPORTS - Use correct config system
@@ -65,9 +65,9 @@ def _create_full_app() -> dash.Dash:
         # Set main layout
         app.layout = _create_main_layout()
 
-        # Register all callbacks using UnifiedCallbackCoordinator
-        coordinator = UnifiedCallbackCoordinator(app)
-        _register_global_callbacks(coordinator)
+        # Register all callbacks using CallbackManager
+        callback_manager = CallbackManager(app)
+        _register_global_callbacks(callback_manager)
 
         # Register page/component callbacks
         try:
@@ -76,13 +76,10 @@ def _create_full_app() -> dash.Dash:
             from components.device_verification import register_callbacks as register_device_verification
             from pages.deep_analytics.callbacks import register_callbacks as register_deep_callbacks
 
-            register_upload_callbacks(coordinator)
-            register_simple_mapping(coordinator)
-            register_device_verification(coordinator)
-            register_deep_callbacks(coordinator)
-
-            if config_manager.get_app_config().environment == "development":
-                coordinator.print_callback_summary()
+            register_upload_callbacks(callback_manager)
+            register_simple_mapping(callback_manager)
+            register_device_verification(callback_manager)
+            register_deep_callbacks(callback_manager)
         except Exception as e:
             logger.warning(f"Failed to register module callbacks: {e}")
 
@@ -409,7 +406,7 @@ def _get_upload_page() -> Any:
         )
 
 
-def _register_global_callbacks(manager: UnifiedCallbackCoordinator) -> None:
+def _register_global_callbacks(manager: CallbackManager) -> None:
     """Register global application callbacks"""
 
     @manager.callback(
